requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.2
django-allauth==65.1.0
django-crispy-forms==2.3
pillow==11.0.0
sqlparse==0.5.1
